refactor: rename directory NewProfessor on components to PublicProfessor
