+++ src/components/PublicProfessor/Publications/PublicationItem.tsx
