+++ src/components/PublicProfessor/Publications/PublicationsList.tsx
